Move all the css over to stitches
