+++ lib/Components/Card.tsx
@@ -1,0 +1,8 @@
+import { styled } from "@stitches/react"
+
+export const Card = styled("div", {
+  borderRadius: 4,
+  padding: 16,
+  border: "1px solid #DDD",
+  background: "white",
+})
