+++ lib/Components/nav.ts
@@ -1,0 +1,13 @@
+import { styled } from "@stitches/react"
+
+export const NavList = styled("ul", {
+  marginTop: 12,
+  marginLeft: 12,
+  marginBottom: 12,
+})
+
+export const NavHeading = styled("span", { fontWeight: 600, padding: "10px 0" })
+
+export const NavLink = styled("a", { color: "#444" })
+
+export const NavItem = styled("li", { whiteSpace: "nowrap", color: "#444" })
