+++ lib/NotFound.tsx
@@ -1,6 +1,6 @@
 import React from "react"
-import { nameFromPath } from "./helpers"
-import { useComponents } from "./components"
+import { nameFromPath } from "@/helpers"
+import { useComponents } from "@/ComponentContext"
 
 type NotFoundProps = {
   path: string
@@ -9,20 +9,20 @@ type NotFoundProps = {
 
 export const NotFound = ({ path, availablePaths }: NotFoundProps) => {
   const componentName = nameFromPath(path)
-  const { PageHeading, Code, Pre, MainColumn } = useComponents()
+  const { PageHeading, MainColumn } = useComponents()
 
   return (
     <MainColumn>
       <PageHeading>404 - Not Found</PageHeading>
 
       <p>
-        None of the docs you passed to <Code>&lt;DocsApp&gt;</Code> have the
-        path <Code>&quot;{path}&quot;</Code>
+        None of the docs you passed to <code>&lt;DocsApp&gt;</code> have the
+        path <code>&quot;{path}&quot;</code>
       </p>
 
       <p>Try:</p>
 
-      <Pre>{`// path/to/${componentName}.doc.tsx
+      <pre>{`// path/to/${componentName}.doc.tsx
 import { Doc, Demo } from "codedocs"
 import { ${componentName} } from "."
 
@@ -37,13 +37,13 @@ export const SomeDemoScenario = (
     <${componentName} ... />
   </Demo>
 )
-`}</Pre>
+`}</pre>
 
       <p>Here are all the paths which were found:</p>
       <ul>
         {availablePaths.map((path) => (
           <li key={path}>
-            <Code>&quot;{path}&quot;</Code>
+            <code>&quot;{path}&quot;</code>
           </li>
         ))}
       </ul>
