--- lib/components.tsx
@@ -1,2 +1,0 @@
-export * as Defaults from "./Defaults"
-export * from "./ComponentContext"
